+++ Doom3/Source/Core/Macros/Profiling.h
@@ -39,7 +39,7 @@ CONCAT(PROFILING_TAG, _EX_TICK_TIME) = CONCAT(PROFILING_TAG, _END);													
 CONCAT(PROFILING_TAG, _EXECUTE_COUNT)++;																																												\
 if (CONCAT(PROFILING_TAG, _ELAPSED_TIME).count() >= 1000.0)																																								\
 {																																																						\
-dooms::profiling::profilingManager::AddProfilingData(MAKE_STRING(PROFILING_TAG), CONCAT(PROFILING_TAG, _PROFILE_ELAPSED_TIME).count() / static_cast<FLOAT64>(CONCAT(PROFILING_TAG, _EXECUTE_COUNT)));					\
+dooms::profiling::profilingManager::AddProfilingData(MAKE_STRING(PROFILING_TAG), static_cast<FLOAT32>(CONCAT(PROFILING_TAG, _PROFILE_ELAPSED_TIME).count()) / static_cast<FLOAT32>(CONCAT(PROFILING_TAG, _EXECUTE_COUNT)));					\
 CONCAT(PROFILING_TAG, _ELAPSED_TIME) = std::chrono::duration<FLOAT64, std::milli>::zero();																																\
 CONCAT(PROFILING_TAG, _PROFILE_ELAPSED_TIME) = std::chrono::duration<FLOAT64, std::milli>::zero();																														\
 CONCAT(PROFILING_TAG, _EXECUTE_COUNT) = 0;																																												\
