+++ Doom3/Source/Core/Reflection/ReflectionType/DEnum.h
@@ -33,11 +33,11 @@ namespace dooms
 			/// <param name="primitiveNameType"></param>
 			/// <returns></returns>
 			const char* GetNameOfEnumConstantValue(const INT32 value, const dooms::reflection::ePrimitiveNameType primitiveNameType = dooms::reflection::ePrimitiveNameType::Short) const;
-			const char* GetNameOfEnumConstantIndex(const INT32 index, const dooms::reflection::ePrimitiveNameType primitiveNameType = dooms::reflection::ePrimitiveNameType::Short) const;
+			const char* GetNameOfEnumConstantIndex(const size_t index, const dooms::reflection::ePrimitiveNameType primitiveNameType = dooms::reflection::ePrimitiveNameType::Short) const;
 
 			// if value is found, return true
 			const bool GetEnumConstantValue(const char* const valueName, INT32& result) const;
-			const bool GetEnumConstantValue(const INT32 index, INT32& result) const;
+			const bool GetEnumConstantValue(const size_t index, INT32& result) const;
 
 			size_t GetEnumConstantIndex(const char* const enumConstantName) const;
 			size_t GetEnumConstantIndex(const INT32 value) const;
